backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
